requirements.txt
bidict==0.22.1
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
flask==3.0.0
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
h11==0.14.0
idna==3.6
importlib-metadata==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pynput==1.7.6
python-engineio==4.8.0
python-socketio==5.10.0
requests==2.31.0
simple-websocket==1.0.0
six==1.16.0
urllib3==2.1.0
werkzeug==3.0.1
wsproto==1.2.0
zipp==3.17.0
